agents/compliance_agent/requirements.txt
nats-py>=2.0.0
python-dotenv>=1.0.0
opa>=1.0.0